+++ src/Bitrix24Partners/Console/ScrapePartnersCommandV2.php
@@ -4,15 +4,11 @@ declare(strict_types=1);
 
 namespace Bitrix24\Lib\Bitrix24Partners\Console;
 
-use Http\Discovery\Psr17FactoryDiscovery;
-use Http\Discovery\Psr18ClientDiscovery;
+use Bitrix24\Lib\Bitrix24Partners\Infrastructure\Scraper\PartnerHtmlParser;
+use Bitrix24\Lib\Bitrix24Partners\Infrastructure\Scraper\PartnerPageScraper;
 use League\Csv\Reader;
 use League\Csv\Writer;
-use Psr\Http\Client\ClientInterface;
-use Psr\Http\Message\RequestFactoryInterface;
-use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Log\LoggerInterface;
-use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\ProgressBar;
@@ -31,9 +27,7 @@ class ScrapePartnersCommandV2 extends Command
     private const DEFAULT_OUTPUT_FILE = 'partners.csv';
     private const DEFAULT_PAGE_DELAY = 3;
     private const DEFAULT_PARTNER_DELAY = 2;
-    private const HTTP_TIMEOUT = 10;
     private const PARTNERS_PER_PAGE = 16;
-    private const BINARY_SEARCH_STEP = 100;
     private const CSV_HEADERS = [
         'bitrix24_partner_number',
         'title',
@@ -47,15 +41,12 @@ class ScrapePartnersCommandV2 extends Command
         'scraped_at',
     ];
 
-    private ?ClientInterface $httpClient = null;
-    private ?RequestFactoryInterface $requestFactory = null;
-    private ?StreamFactoryInterface $streamFactory = null;
-    private LoggerInterface $logger;
-
-    public function __construct(LoggerInterface $logger)
-    {
+    public function __construct(
+        private readonly LoggerInterface $logger,
+        private readonly PartnerPageScraper $scraper,
+        private readonly PartnerHtmlParser $parser,
+    ) {
         parent::__construct();
-        $this->logger = $logger;
     }
 
     protected function configure(): void
@@ -147,7 +138,6 @@ class ScrapePartnersCommandV2 extends Command
             return Command::FAILURE;
         }
 
-        $state = null;
         $startPage = 1;
         $lastPage = 0;
         $processedNumbers = [];
@@ -172,7 +162,7 @@ class ScrapePartnersCommandV2 extends Command
             ));
         } else {
             $io->section('Определение количества страниц...');
-            $lastPage = $this->findLastPage($baseUrl, $insecure, $io);
+            $lastPage = $this->scraper->findLastPage($baseUrl, $insecure, $io);
 
             $io->success(sprintf('Найдено страниц: %d (≈%d партнёров)', $lastPage, $lastPage * self::PARTNERS_PER_PAGE));
         }
@@ -207,13 +197,13 @@ class ScrapePartnersCommandV2 extends Command
             $progressBar->setMessage((string) $page, 'page');
 
             try {
-                $html = $this->fetchPageHtml($page, $baseUrl, $insecure);
+                $html = $this->scraper->fetchPageHtml($page, $baseUrl, $insecure);
                 if ($html === null) {
                     $this->logger->warning(sprintf('Страница %d пустая, пропускаем', $page));
                     continue;
                 }
 
-                $partners = $this->parsePartnerListPage($html);
+                $partners = $this->parser->parsePartnerListPage($html);
             } catch (\Throwable $e) {
                 $this->logger->error(sprintf('Ошибка при обработке страницы %d: %s', $page, $e->getMessage()));
                 continue;
@@ -229,7 +219,7 @@ class ScrapePartnersCommandV2 extends Command
                 }
 
                 try {
-                    $detailHtml = $this->fetchPartnerDetailHtml($partner['detail_page_url'], $insecure);
+                    $detailHtml = $this->scraper->fetchPartnerDetailHtml($partner['detail_page_url'], $insecure);
 
                     $detailData = [
                         'phone' => '',
@@ -238,7 +228,7 @@ class ScrapePartnersCommandV2 extends Command
                         'site' => '',
                     ];
                     if ($detailHtml !== null) {
-                        $detailData = $this->parsePartnerDetailPage($detailHtml);
+                        $detailData = $this->parser->parsePartnerDetailPage($detailHtml);
                     }
 
                     $this->writePartnerToCsv($csvWriter, array_merge($partner, $detailData));
@@ -328,14 +318,14 @@ class ScrapePartnersCommandV2 extends Command
             $io->text(sprintf('Обновление партнёра #%d...', $partnerId));
 
             try {
-                $detailHtml = $this->fetchPartnerDetailHtml($detailPageUrl, $insecure);
+                $detailHtml = $this->scraper->fetchPartnerDetailHtml($detailPageUrl, $insecure);
                 if ($detailHtml === null) {
                     $io->warning(sprintf('Партнёр #%d: не удалось загрузить детальную страницу', $partnerId));
                     ++$errors;
                     continue;
                 }
 
-                $detailData = $this->parsePartnerDetailPage($detailHtml);
+                $detailData = $this->parser->parsePartnerDetailPage($detailHtml);
 
                 $records[$partnerId]['phone'] = $detailData['phone'];
                 $records[$partnerId]['email'] = $detailData['email'];
@@ -395,218 +385,17 @@ class ScrapePartnersCommandV2 extends Command
         return $this->executePartnerUpdate($partnerIds, $outputFile, $partnerDelay, $insecure, $io);
     }
 
-    private function findLastPage(string $baseUrl, bool $insecure, SymfonyStyle $io): int
-    {
-        $io->text('Проверяем страницу 1...');
-        if (!$this->hasPartnersOnPage(1, $baseUrl, $insecure)) {
-            throw new \RuntimeException('Страница 1 не существует или не содержит партнёров. Проверьте URL и доступность сайта.');
-        }
-        $io->text('Страница 1 существует');
-
-        $low = 1;
-        $high = self::BINARY_SEARCH_STEP;
-
-        while ($this->hasPartnersOnPage($high, $baseUrl, $insecure)) {
-            $io->text(sprintf('Страница %d существует, проверяем %d...', $high, $high + self::BINARY_SEARCH_STEP));
-            $low = $high;
-            $high += self::BINARY_SEARCH_STEP;
-            sleep(1);
-        }
-
-        $io->text(sprintf('Страница %d не существует, бинарный поиск между %d и %d...', $high, $low, $high));
-
-        while ($low < $high - 1) {
-            $mid = (int) ceil(($low + $high) / 2);
-
-            if ($this->hasPartnersOnPage($mid, $baseUrl, $insecure)) {
-                $io->text(sprintf('Страница %d существует', $mid));
-                $low = $mid;
-            } else {
-                $io->text(sprintf('Страница %d не существует', $mid));
-                $high = $mid;
-            }
-            sleep(1);
-        }
-
-        return $low;
-    }
-
-    private function hasPartnersOnPage(int $pageNumber, string $baseUrl, bool $insecure): bool
-    {
-        $html = $this->fetchPageHtml($pageNumber, $baseUrl, $insecure);
-        if ($html === null) {
-            return false;
-        }
-
-        $crawler = new Crawler($html);
-
-        return $crawler->filter('div.bp-partner-list-item-cnr.js-partners-list-item')->count() > 0;
-    }
-
-    private function fetchPageHtml(int $pageNumber, string $baseUrl, bool $insecure): ?string
-    {
-        $request = $this->getRequestFactory()->createRequest('POST', $baseUrl)
-            ->withHeader('User-Agent', 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36')
-            ->withHeader('X-Requested-With', 'XMLHttpRequest')
-            ->withHeader('Content-Type', 'application/x-www-form-urlencoded')
-            ->withHeader('Referer', 'https://www.bitrix24.kz/partners/');
-
-        $body = $this->getStreamFactory()->createStream(
-            http_build_query(['ajax' => 'Y', 'page_n' => $pageNumber])
-        );
-        $request = $request->withBody($body);
-
-        $response = $this->getHttpClient($insecure)->sendRequest($request);
-        $content = $response->getBody()->getContents();
-
-        $data = json_decode($content, true);
-        if (!is_array($data) || empty($data['html'])) {
-            return null;
-        }
-
-        return $data['html'];
-    }
-
-    /**
-     * @return array<int, array{partner_number: int, title: string, detail_page_url: string, phone: string}>
-     */
-    private function parsePartnerListPage(string $html): array
-    {
-        $crawler = new Crawler($html);
-        $partners = [];
-
-        $crawler->filter('div.bp-partner-list-item-cnr.js-partners-list-item')->each(
-            function (Crawler $node) use (&$partners) {
-                $partnerNumber = (int) $node->attr('data-partner-id');
-                if ($partnerNumber === 0) {
-                    return;
-                }
-
-                $title = '';
-                $detailPageUrl = '';
-                $nameLink = $node->filter('a.bp-partner-list-item-name')->first();
-                if ($nameLink->count() > 0) {
-                    $title = trim($nameLink->text());
-                    $detailPageUrl = $nameLink->attr('href') ?? '';
-                }
-
-                $phone = '';
-                $phoneNode = $node->filter('div.bp-partner-request-phone')->first();
-                if ($phoneNode->count() > 0) {
-                    $phone = trim($phoneNode->text());
-                }
-
-                $partners[$partnerNumber] = [
-                    'partner_number' => $partnerNumber,
-                    'title' => $title,
-                    'detail_page_url' => $detailPageUrl,
-                    'phone' => $phone,
-                ];
-            }
-        );
-
-        return $partners;
-    }
-
-    private function fetchPartnerDetailHtml(string $detailPageUrl, bool $insecure): ?string
-    {
-        if ($detailPageUrl === '') {
-            return null;
-        }
-
-        $fullUrl = 'https://www.bitrix24.kz' . $detailPageUrl;
-
-        try {
-            $request = $this->getRequestFactory()->createRequest('GET', $fullUrl)
-                ->withHeader('User-Agent', 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36')
-                ->withHeader('Referer', 'https://www.bitrix24.kz/partners/');
-
-            $response = $this->getHttpClient($insecure)->sendRequest($request);
-
-            return $response->getBody()->getContents();
-        } catch (\Throwable $e) {
-            $this->logger->warning(sprintf('Ошибка при загрузке детальной страницы %s: %s', $fullUrl, $e->getMessage()));
-
-            return null;
-        }
-    }
-
-    /**
-     * @return array{email: string, logo_url: string, site: string, phone: string}
-     */
-    private function parsePartnerDetailPage(string $html): array
-    {
-        $crawler = new Crawler($html);
-
-        $phone = '';
-        try {
-            $contactsNode = $crawler->filter('div.bx-partner-detail-description-contacts-content')->first();
-            if ($contactsNode->count() > 0) {
-                $contactsNode->filter('p')->each(function (Crawler $p) use (&$phone) {
-                    $b = $p->filter('b');
-                    if ($b->count() > 0 && str_contains(trim($b->text()), 'Телефон')) {
-                        $fullText = trim($p->text());
-                        $label = trim($b->text());
-                        $extracted = trim(str_replace($label, '', $fullText));
-                        if ($extracted !== '') {
-                            $phone = $extracted;
-                        }
-                    }
-                });
-            }
-        } catch (\Throwable $e) {
-            $this->logger->warning(sprintf('Ошибка парсинга phone: %s', $e->getMessage()));
-        }
-
-        $email = '';
-        try {
-            $contactsNode = $crawler->filter('div.bx-partner-detail-description-contacts-content')->first();
-            if ($contactsNode->count() > 0) {
-                $contactsNode->filter('a')->each(function (Crawler $a) use (&$email) {
-                    $text = trim($a->text());
-                    if (str_contains($text, '@')) {
-                        $email = $text;
-                    }
-                });
-            }
-        } catch (\Throwable $e) {
-            $this->logger->warning(sprintf('Ошибка парсинга email: %s', $e->getMessage()));
-        }
-
-        $logoUrl = '';
-        try {
-            $logoNode = $crawler->filter('img.bx-partner-detail-header-logo-img')->first();
-            if ($logoNode->count() > 0) {
-                $logoUrl = $logoNode->attr('src') ?? '';
-            }
-        } catch (\Throwable $e) {
-            $this->logger->warning(sprintf('Ошибка парсинга logo_url: %s', $e->getMessage()));
-        }
-
-        $site = '';
-        try {
-            $siteNode = $crawler->filter('a.bx-partner-detail-header-info-link')->first();
-            if ($siteNode->count() > 0) {
-                $site = $siteNode->attr('href') ?? '';
-            }
-        } catch (\Throwable $e) {
-            $this->logger->warning(sprintf('Ошибка парсинга site: %s', $e->getMessage()));
-        }
-
-        return [
-            'phone' => $phone,
-            'email' => $email,
-            'logo_url' => $logoUrl,
-            'site' => $site,
-        ];
-    }
-
     private function createCsvWriter(string $outputFile): Writer
     {
         $writer = Writer::from($outputFile, 'w+');
         $writer->insertOne(self::CSV_HEADERS);
 
         return $writer;
+    }
+
+    private function createCsvWriterForResume(string $outputFile): Writer
+    {
+        return Writer::from($outputFile, 'a+');
     }
 
     /**
@@ -697,48 +486,5 @@ class ScrapePartnersCommandV2 extends Command
         }
 
         return $numbers;
-    }
-
-    private function createCsvWriterForResume(string $outputFile): Writer
-    {
-        return Writer::from($outputFile, 'a+');
-    }
-
-    private function getHttpClient(bool $insecure = false): ClientInterface
-    {
-        if ($insecure && $this->httpClient === null) {
-            $symfonyClient = \Symfony\Component\HttpClient\HttpClient::create([
-                'verify_peer' => false,
-                'verify_host' => false,
-                'timeout' => self::HTTP_TIMEOUT,
-            ]);
-            $this->httpClient = new \Symfony\Component\HttpClient\Psr18Client($symfonyClient);
-
-            return $this->httpClient;
-        }
-
-        if ($this->httpClient === null) {
-            $this->httpClient = Psr18ClientDiscovery::find();
-        }
-
-        return $this->httpClient;
-    }
-
-    private function getRequestFactory(): RequestFactoryInterface
-    {
-        if ($this->requestFactory === null) {
-            $this->requestFactory = Psr17FactoryDiscovery::findRequestFactory();
-        }
-
-        return $this->requestFactory;
-    }
-
-    private function getStreamFactory(): StreamFactoryInterface
-    {
-        if ($this->streamFactory === null) {
-            $this->streamFactory = Psr17FactoryDiscovery::findStreamFactory();
-        }
-
-        return $this->streamFactory;
     }
 }
